fix unnecessary create handler func and database func

+++ models/Validators.go
@@ -21,7 +21,7 @@ var addressSchema = bson.M{
 			"description": "must be a string and is required",
 		},
 		"cityCode": bson.M{
-			"bsonType":    "int",
+			"bsonType":    "double",
 			"description": "must be a string and is required",
 		},
 	},
@@ -87,7 +87,7 @@ var customerSchema = bson.M{
 					"description": "must be a string and is required",
 				},
 				"cityCode": bson.M{
-					"bsonType":    "int",
+					"bsonType":    "double",
 					"description": "must be a int and is required",
 				},
 			},
@@ -121,7 +121,7 @@ var orderSchema = bson.M{
 			"description": "must be a string and is required",
 		},
 		"quantity": bson.M{
-			"bsonType":    "int",
+			"bsonType":    "double",
 			"description": "must be a int and is required",
 		},
 		"price": bson.M{
@@ -158,7 +158,7 @@ var orderSchema = bson.M{
 					"description": "must be a string and is required",
 				},
 				"cityCode": bson.M{
-					"bsonType":    "int",
+					"bsonType":    "double",
 					"description": "must be a int and is required",
 				},
 			},
